Changed several data structures associated with rigid contact to use double precision, including C_q, E_q, contact normal, contact plane, friction, cohesion, etc. Updated the APGD iterative solver to match Heyn's PhD thesis.

+++ src/lcp/ChLcpConstraintTwoBodies.cpp
@@ -80,19 +80,19 @@ void ChLcpConstraintTwoBodies::Update_auxiliary()
 	//   the matrices [Eq_a]=[invM_a]*[Cq_a]' and [Eq_b]
 	if (variables_a->IsActive())
 	{
-		ChMatrixNM<float,6,1> mtemp1;
+		ChMatrixNM<double,6,1> mtemp1;
 		mtemp1.CopyFromMatrixT(Cq_a);
 		variables_a->Compute_invMb_v(Eq_a, mtemp1);
 	}
 	if (variables_b->IsActive())
 	{
-		ChMatrixNM<float,6,1> mtemp1;
+		ChMatrixNM<double,6,1> mtemp1;
 		mtemp1.CopyFromMatrixT(Cq_b);
 		variables_b->Compute_invMb_v(Eq_b, mtemp1);
 	}
 
 	//2- Compute g_i = [Cq_i]*[invM_i]*[Cq_i]' + cfm_i 
-	ChMatrixNM<float,1,1> res;
+	ChMatrixNM<double,1,1> res;
 	g_i = 0;
 	if (variables_a->IsActive())
 	{
